code/requirements.txt
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
distro==1.5.0
Django==3.1.7
django-extra-views==0.13.0
djangorestframework==3.12.2
docopt==0.6.2
idna==2.10
importlib-metadata==3.7.0
jsonschema==3.2.0
paramiko==2.7.2
pkg-resources==0.0.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.57.0
zipp==3.4.0
